requirements.txt
# Discord Study Bot - Requirements
# Install with: pip install -r requirements.txt

# Discord
discord.py>=2.0.0
PyNaCl>=1.5.0

# Music
yt-dlp>=2024.0.0

# Configuration
PyYAML>=6.0

# Utilities
pytz>=2023.3
requests>=2.31.0

# AI APIs
openai>=1.0.0
google-generativeai>=0.3.0

# Wolfram
wolframclient>=1.1.7
